documentation for building

+++ src/main/java/pl/put/poznan/transformer/app/Pomieszczenie.java
@@ -34,7 +34,7 @@ public class Pomieszczenie extends Lokalizacja {
         return heating;
     }
 
-    public int getFullLight() {
+    public float getFullLight() {
         return light;
     }
 }
